task-stats-service/requirements.txt
Django>=5.0
djangorestframework>=3.15.0
requests>=2.31.0
tenacity
kafka-python==2.0.2